requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
httpx==0.24.1 